tests: move integration tests to separate modules
