+++ cli/tests/compile_tests.rs
