+++ cli/tests/coverage_tests.rs
