+++ cli/tests/doc_tests.rs
@@ -1,10 +1,10 @@
 // Copyright 2018-2022 the Deno authors. All rights reserved. MIT license.
 
+mod integration;
+
 use test_util as util;
 use test_util::TempDir;
 use util::assert_contains;
-
-use crate::itest;
 
 itest!(deno_doc_builtin {
   args: "doc",
