+++ cli/tests/inspector_tests.rs
