+++ cli/tests/install_tests.rs
