+++ cli/tests/lsp_tests.rs
