+++ cli/tests/repl_tests.rs
