+++ cli/tests/test_tests.rs
@@ -1,6 +1,6 @@
 // Copyright 2018-2022 the Deno authors. All rights reserved. MIT license.
 
-use crate::itest;
+mod integration;
 use deno_core::url::Url;
 use test_util as util;
