+++ cli/tests/vendor_tests.rs
