+++ cli/tests/watcher_tests.rs
